src/use-cases/authenticate.ts
import type { User } from "@prisma/client"
import { compare } from "bcryptjs"

import type { UsersRepository } from "@/repositories/users-repository"

import { InvalidCredentials } from "./errors/invalid-credentials"

interface AuthenticateUseCaseRequest {
  email: string
  password: string
}

interface AuthenticateUseCaseResponse {
  user: User
}

export class AuthenticateUseCase {
  constructor(private usersRepository: UsersRepository) {}

  async execute({
    email,
    password,
  }: AuthenticateUseCaseRequest): Promise<AuthenticateUseCaseResponse> {
    const user = await this.usersRepository.findByEmail(email)

    if (!user) {
      throw new InvalidCredentials()
    }

    const doesPasswordMatches = await compare(password, user.passwordHash)

    if (!doesPasswordMatches) {
      throw new InvalidCredentials()
    }

    return { user }
  }
}
